move new types to `cyclonedx.go`
