--- vulnerability.go
@@ -1,145 +1,0 @@
-// This file is part of CycloneDX Go
-//
-// Licensed under the Apache License, Version 2.0 (the “License”);
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an “AS IS” BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-// Copyright (c) OWASP Foundation. All Rights Reserved.
-
-package cyclonedx
-
-type AffectedVersions struct {
-	Version string              `json:"version,omitempty" xml:"version,omitempty"`
-	Range   string              `json:"range,omitempty" xml:"range,omitempty"`
-	Status  VulnerabilityStatus `json:"status" xml:"status"`
-}
-
-type Affects struct {
-	Ref   string              `json:"ref" xml:"ref"`
-	Range *[]AffectedVersions `json:"range,omitempty" xml:"versions>version,omitempty"`
-}
-
-type Advisory struct {
-	Title string `json:"title,omitempty" xml:"title,omitempty"`
-	URL   string `json:"url" xml:"url"`
-}
-
-type Credits struct {
-	Organizations *[]OrganizationalEntity  `json:"organizations,omitempty" xml:"organizations>organization,omitempty"`
-	Individuals   *[]OrganizationalContact `json:"individuals,omitempty" xml:"individuals>individual,omitempty"`
-}
-
-type Vulnerability struct {
-	BOMRef         string                    `json:"bom-ref,omitempty" xml:"bom-ref,attr,omitempty"`
-	ID             string                    `json:"id" xml:"id"`
-	Source         *Source                   `json:"source,omitempty" xml:"source,omitempty"`
-	References     *[]VulnerabilityReference `json:"references,omitempty" xml:"references>reference,omitempty"`
-	Ratings        *[]VulnerabilityRating    `json:"ratings,omitempty" xml:"ratings>rating,omitempty"`
-	CWEs           *[]int                    `json:"cwes,omitempty" xml:"cwes>cwe,omitempty"`
-	Description    string                    `json:"description,omitempty" xml:"description,omitempty"`
-	Detail         string                    `json:"detail,omitempty" xml:"detail,omitempty"`
-	Recommendation string                    `json:"recommendation,omitempty" xml:"recommendation,omitempty"`
-	Advisories     *[]Advisory               `json:"advisories,omitempty" xml:"advisories>advisory,omitempty"`
-	Created        string                    `json:"created,omitempty" xml:"created,omitempty"`
-	Published      string                    `json:"published,omitempty" xml:"published,omitempty"`
-	Updated        string                    `json:"updated,omitempty" xml:"updated,omitempty"`
-	Credits        *Credits                  `json:"credits,omitempty" xml:"credits,omitempty"`
-	Tools          *[]Tool                   `json:"tools,omitempty" xml:"tools>tool,omitempty"`
-	Analysis       *VulnerabilityAnalysis    `json:"analysis,omitempty" xml:"analysis,omitempty"`
-	Affects        *[]Affects                `json:"affects,omitempty" xml:"affects>target,omitempty"`
-}
-
-type VulnerabilityAnalysis struct {
-	State         ImpactAnalysisState         `json:"state,omitempty" xml:"state,omitempty"`
-	Justification ImpactAnalysisJustification `json:"justification,omitempty" xml:"justification,omitempty"`
-	Response      *[]ImpactAnalysisResponse   `json:"response,omitempty" xml:"responses>response,omitempty"`
-	Detail        string                      `json:"detail,omitempty" xml:"detail,omitempty"`
-}
-
-type VulnerabilityRating struct {
-	Source        *Source       `json:"source,omitempty" xml:"source,omitempty"`
-	Score         float64       `json:"score" xml:"score"`
-	Severity      Severity      `json:"severity,omitempty" xml:"severity,omitempty"`
-	Method        ScoringMethod `json:"method,omitempty" xml:"method,omitempty"`
-	Vector        string        `json:"vector,omitempty" xml:"vector,omitempty"`
-	Justification string        `json:"justification,omitempty" xml:"justification,omitempty"`
-}
-
-type VulnerabilityReference struct {
-	ID     string  `json:"id,omitempty" xml:"id,omitempty"`
-	Source *Source `json:"source,omitempty" xml:"source,omitempty"`
-}
-
-type VulnerabilityStatus string
-
-const (
-	VulnerabilityStatusUnknown     VulnerabilityStatus = "unknown"
-	VulnerabilityStatusAffected    VulnerabilityStatus = "affected"
-	VulnerabilityStatusNotAffected VulnerabilityStatus = "unaffected"
-)
-
-type ScoringMethod string
-
-const (
-	ScoringMethodOther   ScoringMethod = "other"
-	ScoringMethodCVSSv2  ScoringMethod = "CVSSv2"
-	ScoringMethodCVSSv3  ScoringMethod = "CVSSv3"
-	ScoringMethodCVSSv31 ScoringMethod = "CVSSv31"
-	ScoringMethodOWASP   ScoringMethod = "OWASP"
-)
-
-type Severity string
-
-const (
-	SeverityUnknown  Severity = "unknown"
-	SeverityNone     Severity = "none"
-	SeverityInfo     Severity = "info"
-	SeverityLow      Severity = "low"
-	SeverityMedium   Severity = "medium"
-	SeverityHigh     Severity = "high"
-	SeverityCritical Severity = "critical"
-)
-
-type ImpactAnalysisJustification string
-
-const (
-	IAJCodeNotPresent               ImpactAnalysisJustification = "code_not_present"
-	IAJCodeNotReachable             ImpactAnalysisJustification = "code_not_reachable"
-	IAJRequiresConfiguration        ImpactAnalysisJustification = "requires_configuration"
-	IAJRequiresDependency           ImpactAnalysisJustification = "requires_dependency"
-	IAJRequiresEnvironment          ImpactAnalysisJustification = "requires_environment"
-	IAJProtectedByCompiler          ImpactAnalysisJustification = "protected_by_compiler"
-	IAJProtectedAtRuntime           ImpactAnalysisJustification = "protected_at_runtime"
-	IAJProtectedAtPerimeter         ImpactAnalysisJustification = "protected_at_perimeter"
-	IAJProtectedByMitigatingControl ImpactAnalysisJustification = "protected_by_mitigating_control"
-)
-
-type ImpactAnalysisResponse string
-
-const (
-	IARCanNotFix           ImpactAnalysisResponse = "can_not_fix"
-	IARWillNotFix          ImpactAnalysisResponse = "will_not_fix"
-	IARUpdate              ImpactAnalysisResponse = "update"
-	IARRollback            ImpactAnalysisResponse = "rollback"
-	IARWorkaroundAvailable ImpactAnalysisResponse = "workaround_available"
-)
-
-type ImpactAnalysisState string
-
-const (
-	IASResolved             ImpactAnalysisState = "resolved"
-	IASResolvedWithPedigree ImpactAnalysisState = "resolved_with_pedigree"
-	IASExploitable          ImpactAnalysisState = "exploitable"
-	IASInTriage             ImpactAnalysisState = "in_triage"
-	IASFalsePositive        ImpactAnalysisState = "false_positive"
-	IASNotAffected          ImpactAnalysisState = "not_affected"
-)
